+++ R/evaluate_parameters.R
@@ -9,20 +9,20 @@
 #' and \code{\link{evaluate_imputed_values}}.
 #' The possible choices are documented in \code{\link{evaluate_imputed_values}}.
 #'
-#' @param est_pars a vector or matrix of estimated parameters
-#' @param true_pars true parameters, normally a vector or a matrix
+#' @param pars_est a vector or matrix of estimated parameters
+#' @param pars_true true parameters, normally a vector or a matrix
 #'
 #' @export
 #'
 #' @examples
 #' evaluate_parameters(1:4, 2:5, "RMSE")
-evaluate_parameters <- function(est_pars, true_pars, criterion = "RMSE",
+evaluate_parameters <- function(pars_est, pars_true, criterion = "RMSE",
                                 tolerance = sqrt(.Machine$double.eps)) {
-  if (!isTRUE(all.equal(dim(est_pars), dim(true_pars))) ||
-    length(est_pars) != length(true_pars)) {
-    stop("the dimensions of est_pars and true_pars must be equal")
+  if (!isTRUE(all.equal(dim(pars_est), dim(pars_true))) ||
+    length(pars_est) != length(pars_true)) {
+    stop("the dimensions of pars_est and pars_true must be equal")
   }
-  calc_evaluation_criterion(est_pars, true_pars, criterion,
+  calc_evaluation_criterion(pars_est, pars_true, criterion,
     M = NULL,
     tolerance = tolerance
   )