+++ NEWS.md
@@ -1,7 +1,8 @@
 # missMethods (development version)
 
 * improve vignette "Generating missing values" (add connections to Santos et al. (2019))
-* rename of `ds` variables, now all should be named `ds_imp`, `ds_orig` etc.
+* rename of `ds` variables: now all should be named `ds_imp`, `ds_orig` etc.
+* rename of `pars` variables: now all should be named `pars_est` or `pars_true`
 
 # missMethods 0.1.0